feat: more helpful preview error
